feat(Dispatch.Registry): replace hash_ring with libring
libring is a pure Elixir implementation of a hash_ring with no external
dependencies. This makes it easier to use than the hash_ring library as
`hash_ring` requires a NIF.

This commit replaces hash_ring with libring. In order to optimize adding
and removing nodes to the hashring, the state is maintained in a
separate GenServer, which replaces the state of all the hash_rings at
the end of the diff. This ensures less messages are passed when handing
the diff.

There are now two named processes, `Dispatch.Registry` and
`Dispatch.Registry.HashRing`, where HashRing maintains a map of
HashRings.